DYNAMIC HIGHLIGHTER FINAL PROJECT, BABY
It's free! Have fun! I never developed it beyond this point. The biggest flaw is that if you duplicate it, all those duplicates will use the same image as the original, which is no bueno. Also, if you highlight multiple lines of the same length, they too will all use the same image, even if you manually set up different highlighter sequences. so I'd need several images of the SAME length in addition to all the images of DIFFERENT lengths, if I really want each and every highlighter to look different, which is what I originally wanted.